+++ src/shuriken/namespace.clj
@@ -1,6 +1,6 @@
 (ns shuriken.namespace)
 
-(defn class-name [klass]
+(defn- class-name [klass]
   (-> (str klass)
       (clojure.string/replace #"(?:class|interface) " "")
       symbol))
@@ -86,23 +86,3 @@
   [ns & body]
   `(binding [*ns* (the-ns ~ns)]
      (eval (quote (do ~@body)))))
-
-(defmacro once-ns
-  "Ensures a namespace is loaded only once, even after using 'require' or 'use'
-  with :reload or :reload-all. Especially useful for namespaces that monkeypatch
-  other namespaces. SHOULD immediately wrap (ns ...).
-
-  Usage:
-  (require 'shuriken.core) # or shuriken.namespace
-
-  (shuriken.core/once-ns
-    (ns my-namespace)
-
-    (def original-func func)
-
-    (defn func [& args]
-      (swap! counter inc)
-      (apply func args)))"
-  [& body]
-  (when-not (find-ns (-> body first second))
-    `(eval `(do ~@'~body))))