+++ test/shuriken/monkey_patches/syntax_quote_test.clj
@@ -3,18 +3,19 @@
             [shuriken.monkey-patches.syntax-quote]
             [clojure.pprint :refer [pprint]]))
 
+(create-ns 'test)
+
 (deftest test-syntax-quote
-  (testing "syntax-quote macro behavior"
-    (is (= (syntax-quote ~(concat [:a] [:b]))
-           '(:a :b))))
   (testing "` reader macro expansion"
     (is (= `abc
-           (syntax-quote abc))))
+           'shuriken.monkey-patches.syntax-quote-test/abc)))
+  (testing "`` reader macro expansion"
+    (is (= ``abc
+            '(clojure.core/syntax-quoted
+              'shuriken.monkey-patches.syntax-quote-test/abc))))
   (testing "pprint syntax-quote -> ` translation"
     (is (= (with-out-str (pprint ``is))
-           "`clojure.test/is\n"))
-    (is (= (with-out-str (pprint '(syntax-quote is)))
-           "`is\n")))
+           "`clojure.test/is\n")))
   (testing "pprint unquote-splicing -> ~@ translation"
     (is (= (with-out-str (pprint '~@x))
            "~@x\n"))